Try to split large collections of alignments
Basically we can have too many alignments in the Pfam.full or other
files to work with. To deal with this I am trying to split the large
collection of alignments into several smaller ones. This should increase
parallelization and allow the whole process to finish faster.
Additionally, I added the coreutils to the image as that is used to
split.